Replace old Move enum with a 2-byte Move struct

+++ src/aws/mod.rs
@@ -27,7 +27,7 @@ pub struct Event {
 pub struct Output {
     pub pv: Vec<String>,
     pub score: f32,
-    pub nodes: u64,
+    pub nodes: u32,
     pub mem_usage: u64,
     pub time_taken: time::Duration,
 }
